Added capability to modify user roles and restrict delete function

+++ convex/files.ts
@@ -34,7 +34,7 @@ async function hasAccessToOrg (ctx: QueryCtx | MutationCtx, orgId: string) {
     }
 
     const hasAccess = 
-            user.orgIds.includes(orgId) ||
+            user.orgIds.some(item => item.orgId === orgId) ||
             user.tokenIdentifier.includes(orgId);
 
     if (!hasAccess) {
@@ -127,6 +127,14 @@ export const deleteFile = mutation({
 
         if (!access) {
             throw new ConvexError('You do not have access to delete this file!');
+        }
+
+        const isAdmin = 
+            access.user.orgIds.find(org => org.orgId === access.file.orgId)
+            ?.role === "admin";
+        
+        if (!isAdmin) {
+            throw new ConvexError('Admin permission is required to delete this file!');
         }
 
         await ctx.db.delete(args.fileId);
